move spat graph back
